requirements.txt
requests
fastapi
pydantic==1.10.10
python-jose
requests_oauthlib
gspread
boto3
botocore
pre-commit
stripe
uvicorn[standard]
passwordgenerator
python-terraform
airpress
cffi
openstacksdk
asyncio
pyopenssl==22.1.0
black==23.9.1
jinja2
commonmark
redis
